+++ board-room/src/main/java/net/stzups/board/data/objects/Document.java
@@ -1,13 +1,8 @@
 package net.stzups.board.data.objects;
 
+import io.netty.buffer.ByteBuf;
 import net.stzups.board.BoardRoom;
-import net.stzups.board.data.objects.canvas.Point;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import net.stzups.board.data.objects.canvas.Canvas;
 
 public class Document {
     private static final String DEFAULT_DOCUMENT_NAME = "Untitled Document";
@@ -16,19 +11,27 @@ public class Document {
     private User owner;
     private String name;
     private String inviteCode;
-    private Map<User, List<Point>> points = new HashMap<>();
+    private Canvas canvas;
 
+    /**
+     * New document
+     */
     public Document(User owner) {
         this.id = BoardRoom.getSecureRandom().nextLong();
         this.owner = owner;
         owner.getOwnedDocuments().add(id);
         this.name = DEFAULT_DOCUMENT_NAME;
+        this.canvas = new Canvas();
     }
 
-    public Document(long id, User owner, String name) {
+    /**
+     * Serialize document from db
+     */
+    public Document(long id, User owner, String name, ByteBuf byteBuf) {
         this.id = id;
         this.owner = owner;
         this.name = name;
+        this.canvas = new Canvas(byteBuf);
     }
 
     public long getId() {
@@ -43,17 +46,8 @@ public class Document {
         return owner;
     }
 
-    public Map<User, List<Point>> getPoints() {
-        return points;
-    }
-
-    public void addPoints(User user, Point[] points) {
-        List<Point> pts = this.points.get(user);
-        if (pts == null) {
-            pts = new ArrayList<>();
-        }
-        pts.addAll(Arrays.asList(points));
-        this.points.put(user, pts);
+    public Canvas getCanvas() {
+        return canvas;
     }
 
     @Override