+++ board-room/src/main/java/net/stzups/board/data/objects/canvas/CanvasObjects.java
@@ -1,7 +1,0 @@
-package net.stzups.board.data.objects.canvas;
-
-import java.util.ArrayList;
-
-public class CanvasObjects<T extends CanvasObject> extends ArrayList<T> {
-
-}