+++ board-room/src/main/java/net/stzups/board/server/websocket/MessageHandler.java
@@ -8,12 +8,11 @@ import net.stzups.board.data.objects.User;
 import net.stzups.board.data.objects.UserSession;
 import net.stzups.board.server.websocket.protocol.client.ClientMessage;
 import net.stzups.board.server.websocket.protocol.client.messages.ClientMessageCreateDocument;
-import net.stzups.board.server.websocket.protocol.client.messages.ClientMessageDraw;
 import net.stzups.board.server.websocket.protocol.client.messages.ClientMessageHandshake;
 import net.stzups.board.server.websocket.protocol.client.messages.ClientMessageOpenDocument;
+import net.stzups.board.server.websocket.protocol.client.messages.ClientMessageUpdateCanvas;
 import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageAddDocument;
 import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageAddUser;
-import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageDrawClient;
 import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageHandshake;
 
 import java.util.HashMap;
@@ -40,10 +39,8 @@ public class MessageHandler extends SimpleChannelInboundHandler<ClientMessage> {
     @Override
     protected void channelRead0(ChannelHandlerContext ctx, ClientMessage message) {
         switch (message.getMessageType()) {
-            case DRAW: {
-                ClientMessageDraw clientMessageDraw = (ClientMessageDraw) message;
-                room.getDocument().addPoints(client.getUser(), clientMessageDraw.getPoints());
-                room.queueMessageExcept(new ServerMessageDrawClient(client, clientMessageDraw.getPoints()), client);//todo this has tons of latency
+            case UPDATE_CANVAS: {
+                room.updateClient(client, ((ClientMessageUpdateCanvas) message).getCanvasObjects());
                 break;
             }
             case OPEN_DOCUMENT: {
@@ -106,7 +103,7 @@ public class MessageHandler extends SimpleChannelInboundHandler<ClientMessage> {
                         client.queueMessage(new ServerMessageAddDocument(BoardRoom.getDatabase().getDocument(id)));
                     }
                 }
-                client.sendMessages();
+                client.flushMessages();
 
                 break;
             }