+++ board-room/src/main/java/net/stzups/board/server/websocket/Room.java
@@ -2,14 +2,21 @@ package net.stzups.board.server.websocket;
 
 import net.stzups.board.BoardRoom;
 import net.stzups.board.data.objects.Document;
+import net.stzups.board.data.objects.canvas.Canvas;
+import net.stzups.board.data.objects.canvas.object.CanvasObject;
+import net.stzups.board.data.objects.canvas.object.CanvasObjectType;
+import net.stzups.board.data.objects.canvas.object.CanvasObjectWrapper;
 import net.stzups.board.server.websocket.protocol.server.ServerMessage;
 import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageAddClient;
 import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageOpenDocument;
 import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageRemoveClient;
+import net.stzups.board.server.websocket.protocol.server.messages.ServerMessageUpdateCanvas;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.Timer;
 import java.util.TimerTask;
@@ -18,14 +25,12 @@ class Room {
     private static final int SEND_PERIOD = 1000;
 
     private static List<Room> rooms = new ArrayList<>();
-    static {//todo send some packets instantly and refactor to somewhere?
+    static {
         new Timer().scheduleAtFixedRate(new TimerTask() {
             @Override
             public void run() {
                 for (Room room : rooms) {
-                    for (Client client : room.clients) {
-                        client.sendMessages();
-                    }
+                    room.update();
                 }
             }
         }, 0, SEND_PERIOD);
@@ -53,6 +58,12 @@ class Room {
         return document;
     }
 
+    void updateClient(Client client, Map<CanvasObjectType, Map<Short, CanvasObjectWrapper>> canvasObjects) {
+        document.getCanvas().update(canvasObjects);
+        ServerMessageUpdateCanvas serverMessageUpdateCanvas = new ServerMessageUpdateCanvas(canvasObjects);
+        queueMessageExcept(serverMessageUpdateCanvas, client);
+    }
+
     /**
      * Creates a new client using its channel.
      * todo
@@ -60,8 +71,9 @@ class Room {
     void addClient(Client client) {
 
         //for the new client
-        client.sendMessage(new ServerMessageOpenDocument(document));
+        //client.sendMessage(new ServerMessageOpenDocument(document));todo remove
         //for the existing clients
+        client.sendMessage(new ServerMessageOpenDocument(document));
         sendMessage(new ServerMessageAddClient(client));
         clients.add(client);
         BoardRoom.getLogger().info("Added " + client + " to " + this);
@@ -115,6 +127,16 @@ class Room {
         for (Client client : clients) {
             client.queueMessage(serverMessage);
         }
+    }
+
+    void flushMessages() {
+        for (Client client : clients) {
+            client.flushMessages();
+        }
+    }
+
+    private void update() {
+        flushMessages();
     }
 
     @Override