+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/MessageEncoder.java
@@ -16,6 +16,7 @@ import java.util.List;
 public class MessageEncoder extends MessageToByteEncoder<List<ServerMessage>> {
     @Override
     protected void encode(ChannelHandlerContext ctx, List<ServerMessage> serverMessages, ByteBuf b) {
+        System.out.println("encoding " + serverMessages.size());
         BinaryWebSocketFrame binaryWebSocketFrame = new BinaryWebSocketFrame();
         ByteBuf byteBuf = binaryWebSocketFrame.content();
         for (ServerMessage serverMessage : serverMessages) {