+++ board-room/src/main/java/net/stzups/board/server/websocket/protocol/server/ServerMessage.java
@@ -2,7 +2,9 @@ package net.stzups.board.server.websocket.protocol.server;
 
 import io.netty.buffer.ByteBuf;
 
+import java.io.BufferedReader;
 import java.nio.charset.StandardCharsets;
+import java.util.Scanner;
 
 /**
  * Represents a packet sent by the server