+++ frontend/html/scripts/Document.js
@@ -1,11 +1,17 @@
-export const canvas = document.getElementById('canvas');
-export const ctx = canvas.getContext('2d');
-
-import LocalClient from './LocalClient.js';
+import {canvas, Canvas, ctx} from "./canvas/Canvas.js";
 import SidebarItem from './SidebarItem.js';
 import Client from './Client.js'
-import socket from './WebSocketHandler.js'
+import socket from './protocol/WebSocketHandler.js'
 import * as User from "./User.js";
+import ClientMessageOpenDocument from "./protocol/client/messages/ClientMessageOpenDocument.js";
+import ClientMessageCreateDocument from "./protocol/client/messages/ClientMessageCreateDocument.js";
+import ClientMessageHandshake from "./protocol/client/messages/ClientMessageHandshake.js";
+import ServerMessageType from "./protocol/server/ServerMessageType.js";
+import WebSocketHandlerType from "./protocol/WebSocketHandlerType.js";
+import ClientMessageUpdateCanvas from "./protocol/client/messages/ClientMessageUpdateCanvas.js";
+import Shape from "./canvas/canvasObjects/Shape.js";
+import {CanvasObjectType} from "./canvas/CanvasObjectType.js";
+import CanvasObjectWrapper from "./canvas/CanvasObjectWrapper.js";
 
 const documents = new Map();
 let activeDocument = null;
@@ -19,13 +25,17 @@ class Document {
         this.sidebarItem = new SidebarItem(this.name, () => {
             if (this.id != null) {
                 if (activeDocument != null) activeDocument.close();
-                socket.sendOpen(this.id)
+
+                activeDocument = this;
+                activeDocument.open();
+
+                socket.send(new ClientMessageOpenDocument(this.id));
             } else {
                 console.log('id was null', this);
             }
         });
         documents.set(this.id, this);
-        this.points = {};
+        this.canvas = new Canvas();
     }
 
     open() {
@@ -34,32 +44,15 @@ class Document {
         console.log('opened ' + this.name);
         this.sidebarItem.setActive(false);
         //window.history.pushState(document.name, document.title, '/d/' + this.id); todo
-        this.points.forEach((points, id) => {
-            ctx.beginPath();
-            points.forEach((point) => {
-                if (point.dt === 0) {
-                    ctx.stroke();//todo only do this at the end
-                    ctx.moveTo(point.x, point.y);
-                } else {
-                    ctx.lineTo(point.x, point.y);
-                }
-            })
-            ctx.stroke();
-        });
-        this.addClient(localClient);
+        this.canvas.draw(-1);
+        //this.addClient(localClient);
     }
 
     close() {
-        localClient.update();
+        //localClient.update();
         ctx.clearRect(0, 0, canvas.width, canvas.height);//todo a loading screen?
         this.clients.forEach((client) => {
             this.removeClient(client.id);
-        })
-    }
-
-    draw(dt) {
-        this.clients.forEach((client) => {
-            client.draw(dt);
         })
     }
 
@@ -76,7 +69,7 @@ class Document {
 
 document.getElementById('add').addEventListener('click', () => {
     if (activeDocument != null) activeDocument.close();
-    socket.sendCreate();
+    socket.send(new ClientMessageCreateDocument());
 });
 
 window.addEventListener('resize', resizeCanvas);
@@ -86,7 +79,9 @@ function resizeCanvas() {
     canvas.width = rect.width;
     canvas.height = rect.height;
     ctx.putImageData(imageData, 0, 0);
-    //todo redraw?
+    if (activeDocument != null) {
+        activeDocument.canvas.resize();
+    }
 }
 resizeCanvas();
 
@@ -97,67 +92,81 @@ function draw(now) {
     last = now;
 
     if (activeDocument != null) {
-        activeDocument.draw(dt);
+        activeDocument.canvas.draw(dt);
     }
 
     window.requestAnimationFrame(draw);
 }
 window.requestAnimationFrame(draw);
 
-socket.addEventListener('protocol.addclient', (event) => {
+socket.addMessageListener(ServerMessageType.ADD_CLIENT, (event) => {
     let client = new Client(event.id, User.getUser(event.userId));
     activeDocument.addClient(client);
     console.log('Add client ', client);
 });
-socket.addEventListener('protocol.removeclient', (event) => {
+socket.addMessageListener(ServerMessageType.REMOVE_CLIENT, (event) => {
     console.log('Remove client ', activeDocument.removeClient(event.id));
 });
-socket.addEventListener('protocol.draw', (event) => {
-    let client = activeDocument.clients.get(event.id);
-    console.log(activeDocument.clients, event.id);
-    event.points.forEach((point) => {
-        client.points.push(point);
-    });
+socket.addMessageListener(ServerMessageType.UPDATE_CANVAS, (event) => {
+    if (activeDocument != null) {
+        activeDocument.canvas.updateMultiple(event.canvasObjectWrappers);
+    } else {
+        console.warn('oops');
+    }
 });
-socket.addEventListener('protocol.adddocument', (event) => {
+socket.addMessageListener(ServerMessageType.ADD_DOCUMENT, (event) => {
     documents.set(event.id, new Document(event.name, event.id));
 });
-socket.addEventListener('protocol.opendocument', (event) => {
-    if (activeDocument != null) {
-        activeDocument.close();
-    }
-
-    activeDocument = documents.get(event.document.id);
-    Object.assign(activeDocument, event.document);
-    activeDocument.open();
-
-});
-socket.addEventListener('protocol.handshake', (event) => {
+socket.addMessageListener(ServerMessageType.HANDSHAKE, (event) => {
     window.localStorage.setItem('token', event.token.toString());
 })
-socket.addEventListener('socket.open', () => {
+socket.addMessageListener(ServerMessageType.OPEN_DOCUMENT, (event) => {
+    activeDocument.canvas = event.canvas;
+})
+socket.addEventListener(WebSocketHandlerType.OPEN, () => {
     let token = window.localStorage.getItem('token');
     if (token != null) {
         token = BigInt(window.localStorage.getItem('token'));
     } else {
         token = BigInt(0);
     }
-    socket.sendHandshake(token);
+    socket.send(new ClientMessageHandshake(token));
     let invite = document.location.href.substring(document.location.href.lastIndexOf("/") + 1);
     if (invite !== '') {
         try {
             let bigint = BigInt(invite);
-            socket.sendOpen(bigint);
+            //socket.sendOpen(bigint); todo
         } catch(e) {
             console.error('improper invite', invite);
         }
     }
 });
 
-const localClient = new LocalClient();
-
 const inviteButton = document.getElementById("inviteButton");
 
 inviteButton.addEventListener('click', (event) => {
+    console.log('invite');
+})
 
+const MAX_TIME = 2000;
+const UPDATE_INTERVAL = 1000;
+let lastUpdate = 0;
+let updateCanvas = new Canvas();
+setInterval(() => {
+    lastUpdate = window.performance.now();
+    if (updateCanvas.updateCanvasObjects.size > 0) {
+        socket.send(new ClientMessageUpdateCanvas(updateCanvas.updateCanvasObjects));//todo breaks the server when the size is 0
+        updateCanvas.updateCanvasObjects.clear();
+    }
+}, UPDATE_INTERVAL);
+
+function getDt() {
+    return (window.performance.now() - lastUpdate) / MAX_TIME * 255;
+}
+
+canvas.addEventListener('click', (event) => {
+    let shape = Shape.create(event.offsetX, event.offsetY, 50, 50);
+    let id = (Math.random() - 0.5) * 32000;
+    updateCanvas.update(CanvasObjectType.SHAPE, id, CanvasObjectWrapper.create(getDt(), shape));
+    activeDocument.canvas.insert(CanvasObjectType.SHAPE, id, shape);
 })