+++ frontend/html/scripts/protocol/WebSocketHandlerType.js
@@ -1,0 +1,5 @@
+let WebSocketHandlerType;
+export default WebSocketHandlerType = {
+    OPEN:0,
+    CLOSE:1,
+};