+++ frontend/html/scripts/protocol/server/messages/ServerMessageRemoveClient.js
@@ -1,0 +1,8 @@
+import ServerMessage from "../ServerMessage.js";
+
+export default class ServerMessageRemoveClient extends ServerMessage {
+    constructor(reader) {
+        super(reader);
+        this.id = reader.readInt16();
+    }
+}